requirements.txt
elevenlabs==1.5.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
anthropic==0.18.1
python-multipart==0.0.9
httpx==0.28.1
websockets==13.1
twilio==9.3.2
